Refactor date parser coded
