+++ src/test/java/seedu/address/model/person/DateTest.java
@@ -2,12 +2,14 @@ package seedu.address.model.person;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.time.LocalDateTime;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.parser.exceptions.ParseException;
 
 public class DateTest {
     @Test
@@ -59,4 +61,63 @@ public class DateTest {
         assertEquals(initialHashCode, date.hashCode());
     }
 
+    // Valid date parsing test
+    @Test
+    public void parseDateTime_validDate_returnsLocalDateTime() throws ParseException {
+        assertEquals(LocalDateTime.of(2024, 2, 29, 18, 0),
+                Date.parseDateString("29/2/2024 1800"));
+    }
+
+    // Leap year checks
+    @Test
+    public void parseDateTime_february29LeapYear_returnsLocalDateTime() throws ParseException {
+        assertEquals(LocalDateTime.of(2024, 2, 29, 0, 0),
+                Date.parseDateString("29/2/2024 0000"));
+    }
+
+    @Test
+    public void parseDateTime_february29NonLeapYear_throwsParseException() {
+        assertThrows(ParseException.class, () -> Date.parseDateString("29/2/2023 0000"),
+                "Invalid date: February 29 is only valid in leap years.");
+    }
+
+    // Month-end validations
+    @Test
+    public void parseDateTime_invalid31stApril_throwsParseException() {
+        ParseException thrown =
+                assertThrows(ParseException.class, () -> Date.parseDateString("31/4/2024 1200"));
+        assertEquals("Invalid date: APRIL cannot have more than 30 days.", thrown.getMessage());
+    }
+
+    @Test
+    public void parseDateTime_invalid31stJune_throwsParseException() {
+        ParseException thrown =
+                assertThrows(ParseException.class, () -> Date.parseDateString("31/6/2024 1200"));
+        assertEquals("Invalid date: JUNE cannot have more than 30 days.", thrown.getMessage());
+    }
+
+    @Test
+    public void parseDateTime_invalid31stFeb_throwsParseException() {
+        ParseException thrown =
+                assertThrows(ParseException.class, () -> Date.parseDateString("31/2/2024 1200"));
+        assertEquals("Invalid date: FEBRUARY cannot have more than 29 days.", thrown.getMessage());
+    }
+
+    // Invalid formats
+    @Test
+    public void parseDateTime_invalidFormat_throwsParseException() {
+        ParseException thrown =
+                assertThrows(ParseException.class, () -> Date.parseDateString("12-31-2024 1200"));
+        assertEquals("Invalid date format! Please use 'd/M/yyyy HHmm'. "
+                + "For example, '2/12/2024 1800'.", thrown.getMessage());
+    }
+
+
+    @Test
+    public void parseDateTime_outOfRangeDayMonth_throwsParseException() {
+        ParseException thrown =
+                assertThrows(ParseException.class, () -> Date.parseDateString("32/1/2024 1200"));
+        assertEquals("Invalid date or time values! Ensure day, month, hour, "
+                + "and minute ranges are correct.", thrown.getMessage());
+    }
 }
